requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
tqdm
python-dotenv
matplotlib
pandas
numpy
seaborn
statsmodels
scikit-learn